Models/Lung_cancer/requirements.txt
torch==2.6.0
torchvision==0.21.0
scikit-learn==1.6.1
pandas==2.2.3
numpy==1.26.4
pillow==10.0.0
joblib==1.4.2
matplotlib==3.9.0
seaborn==0.13.2
kafka-python==2.0.2
opencv-python-headless==4.10.0.84
python-dotenv
shap
pydicom